+++ lib/pages/events/edit.dart
@@ -3,9 +3,12 @@ import 'package:flutter/services.dart';
 import 'package:intl/intl.dart';
 
 import 'package:thdapp/api.dart';
+import 'package:thdapp/components/background_shapes/CurvedTop.dart';
+import 'package:thdapp/components/buttons/GradBox.dart';
+import 'package:thdapp/components/buttons/SolidButton.dart';
+import 'package:thdapp/components/topbar/TopBar.dart';
 import 'package:thdapp/models/event.dart';
 import 'package:thdapp/pages/events/index.dart';
-import '../custom_widgets.dart';
 
 // HELPER FUNCTIONS
 int daysBetween(DateTime from, DateTime to) {
