+++ lib/pages/events/index.dart
@@ -1,11 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:thdapp/api.dart';
+import 'package:thdapp/components/ErrorDialog.dart';
+import 'package:thdapp/components/background_shapes/CurvedTop.dart';
+import 'package:thdapp/components/buttons/GradBox.dart';
+import 'package:thdapp/components/buttons/SolidButton.dart';
+import 'package:thdapp/components/topbar/TopBar.dart';
 import 'package:thdapp/models/event.dart';
 import 'package:thdapp/pages/events/edit.dart';
 import 'package:intl/intl.dart';
 import 'package:url_launcher/url_launcher.dart';
-import '../custom_widgets.dart';
 import 'package:share/share.dart';
 
 class EventsHomeScreen extends StatefulWidget {
