+++ lib/pages/leaderboard.dart
@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'package:thdapp/components/ErrorDialog.dart';
 import 'package:thdapp/components/background_shapes/CurvedTop.dart';
 import 'package:thdapp/components/buttons/GradBox.dart';
 import 'package:thdapp/components/buttons/SolidButton.dart';
+import 'package:thdapp/components/loading/LoadingOverlay.dart';
 import 'package:thdapp/components/topbar/TopBar.dart';
 import 'package:thdapp/models/lb_entry.dart';
-import 'custom_widgets.dart';
 import '../models/profile.dart';
 import 'package:thdapp/api.dart';
