+++ lib/pages/sponsors.dart
@@ -1,14 +1,13 @@
 import 'package:flutter/material.dart';
+import 'package:thdapp/components/ErrorDialog.dart';
 import 'package:thdapp/components/background_shapes/CurvedTop.dart';
 import 'package:thdapp/components/buttons/GradBox.dart';
 import 'package:thdapp/components/buttons/SolidButton.dart';
 import 'package:thdapp/components/topbar/TopBar.dart';
 import 'package:thdapp/pages/bookmarks.dart';
-import 'custom_widgets.dart';
 import '../models/profile.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:thdapp/api.dart';
-// import 'package:barras/barras.dart';
 import 'profile_page.dart';
 import '../models/discord.dart';
 import 'package:url_launcher/url_launcher.dart';
